requirements.txt
streamlit>=1.39
PyPDF2>=3.0
python-dotenv>=1.0
requests>=2.31
tqdm
pytest
pytest-mock
rich
